requirements.txt
flask
flask-cors
beautifulsoup4
google-generativeai
langchain_google_genai
faiss-cpu
langchain
python-dotenv
langchainhub